JWT Claims

+++ resources/item.py
@@ -1,6 +1,6 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models import ItemModel
 from schemas import ItemSchema, ItemUpdateSchema, StoreSchema
 from db import db
@@ -22,6 +22,9 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin previlage required")
         item = ItemModel.query.get_or_404(item_id)
         
         db.session.delete(item)
